requirements.txt
click~=7.1.2
loguru~=0.5.3
PyYAML~=5.4.1
dotmap~=1.3.23
sqlparse~=0.4.1
pyspark~=2.4.0
pymysql~=1.0.2